Add response creation to edit function on intercepted requests

+++ libmproxy/proxy.py
@@ -279,11 +279,17 @@ class ProxyHandler(SocketServer.StreamRequestHandler):
             if request.kill:
                 self.finish()
                 return
-            server = ServerConnection(request)
-            response = server.read_response()
-            response = response.send(self.mqueue)
+            if request.is_response():
+                response = request
+                request = False
+                response = response.send(self.mqueue)
+            else:
+                server = ServerConnection(request)
+                response = server.read_response()
+                response = response.send(self.mqueue)
+                if response.kill:
+                    server.terminate()
             if response.kill:
-                server.terminate()
                 self.finish()
                 return
             self.send_response(response)
